+++ openbank-notification-service/CHANGELOG.md
@@ -1,5 +1,12 @@
 # Changelog
 
+## [0.29.0](https://github.com/JiRaska/open-bank-oss/compare/notification-service-v0.28.0...notification-service-v0.29.0) (2026-08-27)
+
+
+### Features
+
+* **notification:** add safe no-device push fallback ([#7463](https://github.com/JiRaska/open-bank-oss/issues/7463)) ([7c9cc2c](https://github.com/JiRaska/open-bank-oss/commit/7c9cc2c2b87b8d2a6d1a9fc4a5ef2b94a5935012))
+
 ## [0.28.0](https://github.com/JiRaska/open-bank-oss/compare/notification-service-v0.27.0...notification-service-v0.28.0) (2026-08-24)